move random-number-generator into starknet-devnet-core
