+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
 	"crates/starknet-devnet-server",
 	"crates/starknet-devnet-cli",
 	"crates/starknet-devnet-types",
-	"crates/random-number-generator",
 	"crates/starknet-devnet-core",
 ]
 
@@ -49,6 +48,7 @@ anyhow = "1"
 indexmap = "2.0.0"
 rand = "0.8.5"
 rand_chacha = "0.3.1"
+rand_mt = "4.2.2"
 regex_generate = "0.2.3"
 
 # Starknet dependencies
@@ -70,7 +70,6 @@ starknet-types = { path = "crates/starknet-devnet-types", package = "starknet-de
 starknet-core = { path = "crates/starknet-devnet-core", package = "starknet-devnet-core" }
 server = { path = "crates/starknet-devnet-server", package = "starknet-devnet-server"}
 
-random-number-generator = { path = "crates/random-number-generator", package = "random-number-generator" }
 num-bigint = { version = "0.4" }
 
 lazy_static = { version = "1.4.0" }
