--- crates/random-number-generator/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "random-number-generator"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rand = { version = "0.8.5" }
-rand_mt = "4.2.2"
