+++ crates/starknet-devnet-core/Cargo.toml
@@ -20,7 +20,8 @@ starknet-rs-signers = { workspace = true }
 starknet-rs-ff = { workspace = true }
 starknet-rs-core = { workspace = true }
 starknet-types = { workspace = true }
-random-number-generator = { workspace = true }
+rand = { workspace = true }
+rand_mt = { workspace = true }
 tracing = { workspace = true }
 indexmap = { workspace = true }
